Remove async
While profiling my app I've noticed that a lot of time in `insertScalarBuilder`'s inner function was spent running micro tasks. I haven't benchmarked it, and it's likely that V8 is already optimizing this case, but just in case I've removed a useless `async/await` from `insertTokenScoreParameters` which could be the culprit.